// src/app/Model/movimentacao.ts
export class Movimentacao {
    idMovimentacao?: number;
    idContratante?: string;
    idContratado?: string;
    idPublicacao?: string;
    valorServico?: number;
    horaInicio?: number;
    horaFinal?: number;
    valorPagamento?:number;
    status?: string;
    statusPagamento?: string;
    porcentagemSistema?:number;
    horasTrabalhadas?: number;
    taxaServico?: number;

}